com.spring.dozen.auth/src/main/java/com/spring/dozen/auth/infra/config/jwt/JwtTokenProvider.java
package com.spring.dozen.auth.infra.config.jwt;

import com.spring.dozen.auth.application.interfaces.TokenProvider;
import com.spring.dozen.auth.domain.enums.Role;
import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.io.Decoders;
import io.jsonwebtoken.security.Keys;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

import javax.crypto.SecretKey;
import java.util.Date;

@Component
public class JwtTokenProvider implements TokenProvider {

    @Value("${spring.application.name}")
    private String issuer;

    @Value("${service.jwt.access-expiration}")
    private Long accessExpiration;

    private final SecretKey secretKey;

    public JwtTokenProvider(@Value("${service.jwt.secret-key}") String secretKey) {
        this.secretKey = Keys.hmacShaKeyFor(Decoders.BASE64URL.decode(secretKey));
    }

    @Override
    public String createAccessToken(Long userId, Role role) {
        return Jwts.builder()
                .claim("user_id", userId)
                .claim("role", role.name())
                .issuer(issuer)
                .issuedAt(new Date(System.currentTimeMillis()))
                .expiration(new Date(System.currentTimeMillis() + accessExpiration))
                .signWith(secretKey, io.jsonwebtoken.SignatureAlgorithm.HS512)
                .compact();
    }

    SecretKey getSecretKey() {
        return secretKey;
    }
}
